requirements.txt
openai>=1.3.0
httpx>=0.25.0
python-dotenv>=1.0.0
google-genai>=0.1.0
